Extract data for external plugins from internal plugin file
