+++ src/eth_plugin_interface.h
@@ -6,6 +6,7 @@
 #include "os.h"
 #include "cx.h"
 #include "tokens.h"
+#include "tx_content.h"
 
 /*************************************************************************************************
  * Comments provided in this file are quick reminders on the usage of the plugin interface       *
