+++ src_features/setExternalPlugin/cmd_setExternalPlugin.c
@@ -3,6 +3,7 @@
 #include "public_keys.h"
 #include "eth_plugin_interface.h"
 #include "eth_plugin_internal.h"
+#include "plugin_utils.h"
 #include "common_ui.h"
 #include "os_io_seproxyhal.h"
