+++ src_plugins/erc1155/erc1155_provide_parameters.c
@@ -2,6 +2,7 @@
 
 #include <string.h>
 #include "erc1155_plugin.h"
+#include "plugin_utils.h"
 #include "eth_plugin_internal.h"
 #include "utils.h"
