+++ src_plugins/erc1155/erc1155_ui.c
@@ -2,6 +2,7 @@
 
 #include <string.h>
 #include "erc1155_plugin.h"
+#include "eth_plugin_internal.h"
 #include "eth_plugin_interface.h"
 #include "utils.h"
