+++ src_plugins/erc721/erc721_plugin.c
@@ -1,6 +1,7 @@
 #ifdef HAVE_NFT_SUPPORT
 
 #include <string.h>
+#include "plugin_utils.h"
 #include "erc721_plugin.h"
 #include "eth_plugin_internal.h"
 #include "eth_plugin_interface.h"
