+++ src_plugins/erc721/erc721_ui.c
@@ -2,6 +2,7 @@
 
 #include <string.h>
 #include "erc721_plugin.h"
+#include "eth_plugin_internal.h"
 #include "eth_plugin_interface.h"
 #include "utils.h"
