+++ tools/build_sdk.py
@@ -157,7 +157,7 @@ if __name__ == "__main__":
         "src/tokens.h",
         "src/utils.h",
         "src/tx_content.h",
-        "src/eth_plugin_internal.h",
+        "src/plugin_utils.h",
         "src/nft.h",
         "src/extra_info.h",
         "src/caller_api.h",
@@ -200,7 +200,7 @@ if __name__ == "__main__":
     # extract and merge function bodies
     c_files_to_merge = [
         "src/utils.c",
-        "src/eth_plugin_internal.c",
+        "src/plugin_utils.c",
     ]
     merge_c_files(c_files_to_merge, nodes_to_extract["fn"])
